# Optimize updates with ObservableObject