This includes two optimizations:

1. Removed extra wrapping Views, such as `NodeView`. This manual inlining reduced some View.body evaluations.
2. Add explicit `objectWillChange` properties to `ObservableObject` subclasses. Oddly, the slowest part of a View update was the default `objectWillChange` getter provided by the `ObservableObject` protocol. I'm not sure how they implemented this getter, but adding a simple implementation drastically improved update performance.

I tested these changes on a LiveView with around 200 elements and some moderately complex modifier stacks:

Previously, Instruments reported a microhang of ~400ms anytime an updated happened. The initial app load also took over 4 seconds.

After these optimizations, updates typically take <=100ms and the initial app loads in 2.3 seconds.